feat: some new features and fixes on the controller mapping and logic across the application.
