+++ src/main/java/com/cms/claimmanagement/ClaimManagementApplication.java
@@ -4,7 +4,7 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.web.servlet.config.annotation.EnableWebMvc;
 
-@EnableWebMvc
+
 @SpringBootApplication
 public class ClaimManagementApplication {
