--- src/main/java/com/cms/claimmanagement/insuranceCompany/Claim/Controller/ClaimUpdateData.java
@@ -1,5 +1,0 @@
-package com.cms.claimmanagement.insuranceCompany.Claim.Controller;
-
-public record ClaimUpdateData(String claimId, boolean claimStatus, boolean insuranceCompanyApproval) {
-
-}
