--- src/main/java/com/cms/claimmanagement/insuranceCompany/Claim/Repository/ClaimRepository.java
@@ -1,6 +1,0 @@
-package com.cms.claimmanagement.insuranceCompany.Claim.Repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ClaimRepository extends JpaRepository<ClaimDetails, String> {
-}
