--- src/main/java/com/cms/claimmanagement/insuranceCompany/Policy/Controller/PolicyRequestData.java
@@ -1,18 +1,0 @@
-package com.cms.claimmanagement.insuranceCompany.Policy.Controller;
-
-import jakarta.validation.constraints.NotNull;
-
-import java.time.LocalDate;
-
-public record PolicyRequestData(
-
-        String insuredFirstName,
-        String insuredLastName,
-        LocalDate dateOfInsurance,
-
-        String emailId,
-        @NotNull
-        String vehicleNo,
-        boolean status
-) {
-}
