--- src/main/java/com/cms/claimmanagement/insuranceCompany/Policy/Repository/PolicyRepository.java
@@ -1,13 +1,0 @@
-package com.cms.claimmanagement.insuranceCompany.Policy.Repository;
-
-import com.cms.claimmanagement.insuranceCompany.Policy.Repository.PolicyEntity;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface PolicyRepository extends JpaRepository<PolicyEntity, Long> {
-
-
-
-    Optional findByPolicyNo(String s);
-}
