--- src/main/java/com/cms/claimmanagement/insuranceCompany/Surveyor/Controller/SurveyorRequestData.java
@@ -1,4 +1,0 @@
-package com.cms.claimmanagement.insuranceCompany.Surveyor.Controller;
-
-public class SurveyorRequestData {
-}
