--- src/main/java/com/cms/claimmanagement/insuranceCompany/Surveyor/Repository/SurveyorRepository.java
@@ -1,6 +1,0 @@
-package com.cms.claimmanagement.insuranceCompany.Surveyor.Repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface SurveyorRepository extends JpaRepository<Surveyor, Long> {
-}
