--- src/main/java/com/cms/claimmanagement/insuranceCompany/Surveyor/SurveyReport/SurveyReportDTO.java
@@ -1,6 +1,0 @@
-package com.cms.claimmanagement.insuranceCompany.Surveyor.SurveyReport;
-
-public record SurveyReportDTO(String claimId, String policyNo,
-                              Long labourCharges, Long partsCost, Long policyClass,
-                              Long depreciationCost, Long totalAmount) {
-}
