+++ src/main/java/com/cms/claimmanagement/insuranceCompany/claim/controller/ClaimResponseData.java
@@ -1,0 +1,16 @@
+package com.cms.claimmanagement.insuranceCompany.claim.controller;
+
+
+
+public class ClaimResponseData {
+   private String claimId;
+
+    public String getClaimId() {
+        return claimId;
+    }
+    public void setClaimId(String claimId){
+        this.claimId = claimId;
+}
+
+
+}
