+++ src/main/java/com/cms/claimmanagement/insuranceCompany/claim/controller/ClaimUpdateData.java
@@ -1,0 +1,5 @@
+package com.cms.claimmanagement.insuranceCompany.claim.controller;
+
+public record ClaimUpdateData(String claimId, boolean claimStatus, boolean insuranceCompanyApproval) {
+
+}
