+++ src/main/java/com/cms/claimmanagement/insuranceCompany/claim/repository/ClaimRepository.java
@@ -1,0 +1,6 @@
+package com.cms.claimmanagement.insuranceCompany.claim.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface ClaimRepository extends JpaRepository<ClaimDetails, String> {
+}
