+++ src/main/java/com/cms/claimmanagement/insuranceCompany/policy/Controller/PolicyRequestData.java
@@ -1,0 +1,18 @@
+package com.cms.claimmanagement.insuranceCompany.policy.Controller;
+
+import jakarta.validation.constraints.NotNull;
+
+import java.time.LocalDate;
+
+public record PolicyRequestData(
+
+        String insuredFirstName,
+        String insuredLastName,
+        LocalDate dateOfInsurance,
+
+        String emailId,
+        @NotNull
+        String vehicleNo,
+        boolean status
+) {
+}
