+++ src/main/java/com/cms/claimmanagement/insuranceCompany/policy/Repository/PolicyRepository.java
@@ -1,0 +1,12 @@
+package com.cms.claimmanagement.insuranceCompany.policy.Repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface PolicyRepository extends JpaRepository<PolicyEntity, Long> {
+
+
+
+    Optional findByPolicyNo(String s);
+}
