+++ src/main/java/com/cms/claimmanagement/insuranceCompany/surveyor/controller/SurveyorRequestData.java
@@ -1,0 +1,4 @@
+package com.cms.claimmanagement.insuranceCompany.surveyor.controller;
+
+public class SurveyorRequestData {
+}
