+++ src/main/java/com/cms/claimmanagement/insuranceCompany/surveyor/controller/SurveyorResponseData.java
@@ -1,0 +1,4 @@
+package com.cms.claimmanagement.insuranceCompany.surveyor.controller;
+
+public record SurveyorResponseData(Long surveyorId, String firstName, String lastName, Long estimateLimit) {
+}
