+++ src/main/java/com/cms/claimmanagement/insuranceCompany/surveyor/repository/SurveyorRepository.java
@@ -1,0 +1,6 @@
+package com.cms.claimmanagement.insuranceCompany.surveyor.repository;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface SurveyorRepository extends JpaRepository<Surveyor, Long> {
+}
