+++ src/main/java/com/cms/claimmanagement/insuranceCompany/surveyor/surveyReport/SurveyReportDTO.java
@@ -1,0 +1,6 @@
+package com.cms.claimmanagement.insuranceCompany.surveyor.surveyReport;
+
+public record SurveyReportDTO(String claimId, String policyNo,
+                              Long labourCharges, Long partsCost, Long policyClass,
+                              Long depreciationCost, Long totalAmount) {
+}
